LightProbeGrid: Replace SunLight during bake.
Shares the SunLight swap with LightProbeGridWebGL through LightProbeGridUtils. Bake lights are cached per SunLight so ranged bakes reuse their pipelines instead of recompiling every material.
